fix flaky test card tests (#7792)
* add missing mock

* consolidate test constants into shared files

* some more cleanup

* change mock generation

* fix the other two mocks needed

* fix other mocks

* try again

* dedupe const

* delete unused vars
